js: fix "Change my password" functionality
It was broken because the new password was placed in the "passhash" property of
the employee object, but REST server was expecting it to be in "password".

Since the JS side does not need the "passhash" and "salt" properties at all,
this commit drops them from the employee object prototype.

Signed-off-by: Nathan Cutler <ncutler@suse.com>