config: drop derived distro site config file
Once a site config param is set, it is not possible to override it. This is a
feature of App::CELL and it's not worth the effort to change it.

As a result, we put all our site overrides in ext/WWW_SiteConfig.pm.example and
rely on packagers to install it as /etc/dochazka-www/WWW_SiteConfig.pm.

Signed-off-by: Nathan Cutler <ncutler@suse.com>