js: Always load employee profile from server
Whenever we enter the employee profile, (re-)load it and update the cache.
We cannot rely on the cache when displaying the employee profile, because
we might be returning from an update operation that affects the profile.

Also, implement a "Remove supervisor" option in the empProfile miniMenu and
make good use of it in the functional test.

Signed-off-by: Nathan Cutler <ncutler@suse.com>